finalize network status
