+++ html/pfappserver/root/src/views/Status/network/_composables/useTooltips.js
@@ -1,4 +1,7 @@
 import { computed, ref, toRefs } from '@vue/composition-api'
+import d3 from '@/utils/d3'
+import useColor from './useColor'
+import { useViewBox } from './useSvg'
 
 const getAngleFromCoords = (x1, y1, x2, y2) => {
   const dx = x2 - x1
@@ -8,11 +11,20 @@ const getAngleFromCoords = (x1, y1, x2, y2) => {
   return (360 + theta) % 360
 }
 
-export default (props, config, bounds, viewBox, nodes) => {
+export default (props, config, bounds, viewBox, nodes, links) => {
 
   const {
     dimensions
   } = toRefs(props)
+
+  const {
+    centerX,
+    centerY
+  } = useViewBox(config, dimensions)
+
+  const {
+    color
+  } = useColor(props, config)
 
   const localTooltips = ref([]) // private d3 tooltips
 
@@ -64,10 +76,155 @@ export default (props, config, bounds, viewBox, nodes) => {
     return { style }
   }
 
+  let highlight = ref(false) // mouseOver @ node
+  let highlightNodeId = ref(false) // last highlighted node
+
+  const highlightedLinks = computed(() => {
+    return links.value.filter(link => { link.highlight })
+  })
+
+  const mouseOverNode = node => {
+    const { width, height } = dimensions.value
+    _highlightNodeById(node.id) // highlight node
+    highlight.value = (node.type === 'node') ? color(node) : 'none'
+    // tooltips
+    if (highlightNodeId.value !== node.id) {
+      highlightNodeId.value = node.id
+      const highlightedNodes = nodes.value.filter(node => node.tooltip)
+      localTooltips.value = [
+        ...highlightedNodes.map(node => {
+          const { id, type, properties } = node
+          const { x, y } = coordBounded(node)
+          return JSON.parse(JSON.stringify({ id, type, properties, x, y }))
+        }),
+        ...highlightedNodes.map(node => {
+          const { id } = node
+          const { x: fx, y: fy } = coordBounded(node)
+          return JSON.parse(JSON.stringify({ id: `${id}-fixed`, fx, fy }))
+        })
+      ]
+      // link force from tooltip to node (self)
+      let selfLinks = []
+      highlightedNodes.map(node => {
+        const { id } = node
+        selfLinks.push({ source: id, target: `${id}-fixed` })
+      })
+      // link force from tooltip to other nodes
+      let nodeLinks = []
+      highlightedNodes.map(node => {
+        const { id } = node
+        highlightedNodes.map(other => {
+          const { id: otherId } = other
+          if (otherId !== id) {
+            if (nodeLinks.filter(link =>
+              ([link.source, link.target].includes(id) && [link.source, link.target].includes(`${otherId}-fixed`))
+            ).length === 0) {
+              nodeLinks.push({ source: id, target: `${otherId}-fixed` })
+            }
+          }
+        })
+      })
+      // link force from tooltip to other tooltips
+      let tooltipLinks = []
+      highlightedNodes.map(node => {
+        const { id } = node
+        highlightedNodes.map(other => {
+          const { id: otherId } = other
+          if (otherId !== id) {
+            if (tooltipLinks.filter(link =>
+              ([link.source, link.target].includes(id) && [link.source, link.target].includes(otherId))
+            ).length === 0) {
+              tooltipLinks.push({ source: otherId, target: id })
+            }
+          }
+        })
+      })
+      d3.forceSimulation(localTooltips.value)
+        .alphaDecay(1 - Math.pow(0.001, 1 / 50)) // default: 1 - Math.pow(0.001, 1 / 300)
+        .velocityDecay(0.8) // default: 0.4
+        .force('x', d3.forceX() // force: tooltip w/ center x
+          .x(centerX.value)
+          .strength(0.5)
+        )
+        .force('y', d3.forceY() // force: tooltip w/ center y
+          .y(centerY.value)
+          .strength(0.5)
+        )
+        .force('selfLinks', d3.forceLink(selfLinks) // force: tooltip w/ node
+          .id((d) => d.id)
+          .distance(Math.min(width, height) / 8)
+          .strength(0.5)
+          .iterations(4)
+        )
+        .force('nodeLinks', d3.forceLink(nodeLinks) // force: tooltip w/ other nodes
+          .id((d) => d.id)
+          .distance(Math.min(width, height) / 2)
+          .strength(0.5)
+          .iterations(2)
+        )
+        .force('tooltipLinks', d3.forceLink(tooltipLinks) // force: tooltip w/ other tooltips
+          .id((d) => d.id)
+          .distance(Math.min(width, height) / 2)
+          .strength(0.5)
+          .iterations(8)
+        )
+        .restart()
+    }
+  }
+
+  const mouseOutNode = () => {
+    _unhighlightNodes()
+    _unhighlightLinks()
+    highlight.value = false
+    highlightNodeId.value = false
+    localTooltips.value = []
+  }
+
+  const _unhighlightNodes = () => {
+    nodes.value = nodes.value.map(node => ({ ...node, highlight: false, tooltip: false }))
+  }
+
+  const _unhighlightLinks = () => {
+    links.value = links.value.map(link => ({ ...link, highlight: false }))
+  }
+
+  const _highlightNodeById = id => {
+    _unhighlightNodes()
+    _unhighlightLinks()
+    // highlight all target nodes linked to this source node
+    links.value.forEach((link, index)=> {
+      if (link.source.id === id) {
+        links.value[index].highlight = true // highlight link
+        links.value[index].target.highlight = true // highlight target node
+      }
+    })
+    let sourceIndex = nodes.value.findIndex(node => node.id === id)
+    while (sourceIndex > -1) { // travel to center of tree [ (target|source) -> (target|source) -> ... ]
+      nodes.value[sourceIndex].highlight = true  // highlight node
+      nodes.value[sourceIndex].tooltip = true // show node tooltip
+      const { id: sourceId } = nodes.value[sourceIndex]
+      links.value.forEach((link, index) => {
+        const { target: { id: targetId } = {} } = link
+        if (targetId === sourceId) {
+          links.value[index].highlight = true
+        }
+      })
+      if ('source' in nodes.value[sourceIndex])
+        sourceIndex = nodes.value.findIndex(node => node.id === nodes.value[sourceIndex].source.id)
+      else
+        break
+    }
+  }
+
   return {
     coordBounded,
     localTooltips,
     tooltips,
-    tooltipAnchorAttrs
+    tooltipAnchorAttrs,
+    highlight,
+    highlightNodeId,
+    highlightedLinks,
+    mouseOverNode,
+    mouseOutNode
   }
 }
